Converted the prompt code into a function

+++ src/export_result.py
@@ -7,8 +7,9 @@ import webbrowser
 
 
 class ExportResult:
-    def __init__(self, data):
+    def __init__(self, data, prompt):
         self.data = data
+        self.prompt = prompt
     
     def export_to_json(self, file_name):
         f = open(file_name, 'w+')
@@ -22,7 +23,7 @@ class ExportResult:
         print('Result has has been saved in: ' + file_name + '.zip')    
         q = Questions()
 
-        result = q.query_yes_no('Do you want to upload the result to your profile automatically?')
+        result = q.query_yes_no('Do you want to upload the result to your profile automatically?') if self.prompt else False
         if result:
             response = uploadRepo( file_name + '.zip')
             if response is not None:
@@ -32,4 +33,5 @@ class ExportResult:
                 print('Go to this link in the browser => '+ url)
                 webbrowser.open(url)
 
-        os.remove(file_name)
+        if self.prompt:
+            os.remove(file_name)
